refactor useSignUpData to useSignUp
